+++ source/isaaclab/isaaclab/sensors/frame_transformer/__init__.py
@@ -10,3 +10,5 @@ from .base_frame_transformer_data import BaseFrameTransformerData
 from .frame_transformer import FrameTransformer
 from .frame_transformer_cfg import FrameTransformerCfg, OffsetCfg
 from .frame_transformer_data import FrameTransformerData
+
+__all__ = ["BaseFrameTransformer", "BaseFrameTransformerData", "FrameTransformer", "FrameTransformerCfg", "FrameTransformerData", "OffsetCfg"]