+++ source/isaaclab/isaaclab/sensors/imu/__init__.py
@@ -12,3 +12,5 @@ from .base_imu_data import BaseImuData
 from .imu import Imu
 from .imu_cfg import ImuCfg
 from .imu_data import ImuData
+
+__all__ = ["BaseImu", "BaseImuData", "Imu", "ImuCfg", "ImuData"]