+++ source/isaaclab_physx/isaaclab_physx/sensors/contact_sensor/__init__.py
@@ -7,3 +7,5 @@
 
 from .contact_sensor import ContactSensor
 from .contact_sensor_data import ContactSensorData
+
+__all__ = ["ContactSensor", "ContactSensorData"]