+++ source/isaaclab_physx/isaaclab_physx/sensors/frame_transformer/__init__.py
@@ -7,3 +7,5 @@
 
 from .frame_transformer import FrameTransformer
 from .frame_transformer_data import FrameTransformerData
+
+__all__ = ["FrameTransformer", "FrameTransformerData"]